init mobile ads only if remote config set true

+++ lib/shared/repositories/mobile_ads_repository.dart
@@ -14,11 +14,16 @@ class MobileAdsRepository extends IMobileAdsRepository {
   MobileAdsRepository({
     required FirebaseAnalyticsCacheController firebaseAnalyticsCacheController,
     required IAppLayoutRepository appLayoutRepository,
+    required FirebaseRemoteConfigProvider firebaseRemoteConfigProvider,
   })  : _appLayoutRepository = appLayoutRepository,
-        _firebaseAnalyticsCacheController = firebaseAnalyticsCacheController;
+        _firebaseAnalyticsCacheController = firebaseAnalyticsCacheController,
+        _firebaseRemoteConfigProvider = firebaseRemoteConfigProvider;
   final IAppLayoutRepository _appLayoutRepository;
   final FirebaseAnalyticsCacheController _firebaseAnalyticsCacheController;
+  final FirebaseRemoteConfigProvider _firebaseRemoteConfigProvider;
   AdRequest? request;
+
+  static const showMobileAds = '__show_mobile_ads_key__';
 
   final adUnitId = PlatformEnum.getPlatform.isAndroid
       ? Config.isReleaseMode
@@ -31,6 +36,8 @@ class MobileAdsRepository extends IMobileAdsRepository {
   Future<void> _init() async {
     bool nonPersonalizedAds;
     try {
+      await MobileAds.instance.initialize();
+
       if (PlatformEnum.getPlatform.isIOS) {
         final trackingStatus =
             await AppTrackingTransparency.trackingAuthorizationStatus;
@@ -57,39 +64,45 @@ class MobileAdsRepository extends IMobileAdsRepository {
   Future<Either<SomeFailure, BannerAd?>> loadBannerAd() async {
     return eitherFutureHelper(
       () async {
-        if (request == null) await _init();
-        if (PlatformEnum.getPlatform.isIOS) {}
-        final size = _appLayoutRepository.getScreenSize;
+        await _firebaseRemoteConfigProvider.waitActivated();
+        final show = _firebaseRemoteConfigProvider.getBool(showMobileAds);
+        if (show) {
+          if (request == null) await _init();
+          if (PlatformEnum.getPlatform.isIOS) {}
+          final size = _appLayoutRepository.getScreenSize;
 
-        if (size == null) return const Right(null);
+          if (size == null) return const Right(null);
 
-        final adSize =
-            await AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(
-          size.width.truncate(),
-        );
-        if (adSize == null) return const Right(null);
+          final adSize =
+              await AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(
+            size.width.truncate(),
+          );
+          if (adSize == null) return const Right(null);
 
-        final bannerAd = BannerAd(
-          adUnitId: adUnitId,
-          request: request!,
-          size: adSize,
-          listener: BannerAdListener(
-            // Called when an ad is successfully received.
-            onAdLoaded: (ad) {
-              log('$ad loaded.');
-            },
-            // Called when an ad request failed.
-            onAdFailedToLoad: (ad, err) {
-              log('BannerAd failed to load: $err');
-              // Dispose the ad here to free resources.
-              ad.dispose();
-            },
-          ),
-        );
+          final bannerAd = BannerAd(
+            adUnitId: adUnitId,
+            request: request!,
+            size: adSize,
+            listener: BannerAdListener(
+              // Called when an ad is successfully received.
+              onAdLoaded: (ad) {
+                log('$ad loaded.');
+              },
+              // Called when an ad request failed.
+              onAdFailedToLoad: (ad, err) {
+                log('BannerAd failed to load: $err');
+                // Dispose the ad here to free resources.
+                ad.dispose();
+              },
+            ),
+          );
 
-        await bannerAd.load();
+          await bannerAd.load();
 
-        return Right(bannerAd);
+          return Right(bannerAd);
+        } else {
+          return const Right(null);
+        }
       },
       methodName: 'loadBannerAd',
       className: 'Mobile Ads ${ErrorText.repositoryKey}',
